+++ packages/orq-rc/src/models/operations/updateidentity.ts
@@ -167,7 +167,7 @@ export const UpdateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-07T11:51:10.719Z",
+    "2026-03-08T19:14:48.554Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {